Refactor script import to poll from parent component
