+++ src/containers/AdminScriptImport.jsx
@@ -1,17 +1,13 @@
 import PropTypes from "prop-types";
 import React, { Component } from "react";
-import _ from "lodash";
 import { StyleSheet, css } from "aphrodite";
 
-import gql from "graphql-tag";
-import loadData from "./hoc/load-data";
 import theme from "../styles/theme";
 import CampaignFormSectionHeading from "../components/CampaignFormSectionHeading";
 import TextField from "material-ui/TextField";
 import { ListItem, List } from "material-ui/List";
 import RaisedButton from "material-ui/RaisedButton";
 import ErrorIcon from "material-ui/svg-icons/alert/error";
-import { type } from "os";
 
 const errorIcon = <ErrorIcon color={theme.colors.red} />;
 
@@ -21,50 +17,25 @@ const styles = StyleSheet.create({
   }
 });
 
-// TODO[matteo]: refactor this so it doesn't do its own polling
-export class AdminScriptImport extends Component {
+export default class AdminScriptImport extends Component {
+  static propTypes = {
+    startImport: PropTypes.func,
+    hasPendingJob: PropTypes.bool
+  };
+
   constructor(props) {
     super(props);
     this.state = {};
   }
+
   startImport = async () => {
-    const res = await this.props.mutations.importCampaignScript(
-      this.props.campaignData.campaign.id,
-      this.state.url
-    );
+    const res = await this.props.startImport(this.state.url);
     if (res.errors) {
-      this.setState({ error: res.errors.message, importingScript: false });
-    } else {
-      const jobId = res.data.importCampaignScript;
-      this.setState({ importingScript: true, error: undefined });
-      await this.pollDuringActiveJobs(jobId);
+      this.setState({ error: res.errors.message });
     }
   };
 
   handleUrlChange = (_eventId, newValue) => this.setState({ url: newValue });
-
-  pollDuringActiveJobs = async jobId => {
-    const fetchedPendingJobsData = await this.props.pendingJobsData.refetch();
-    const pendingJobs = fetchedPendingJobsData.data.campaign.pendingJobs;
-    const ourJob = _.find(
-      pendingJobs,
-      pendingJob => pendingJob.id === jobId.toString()
-    );
-    if (!ourJob || ourJob.resultMessage) {
-      this.setState({
-        importingScript: false,
-        error: !!ourJob && ourJob.resultMessage
-      });
-
-      if (!ourJob) {
-        this.props.onSubmit();
-      }
-      return;
-    }
-    setTimeout(async () => {
-      await this.pollDuringActiveJobs(jobId);
-    }, 1000);
-  };
 
   renderErrors = () =>
     this.state.error && (
@@ -90,7 +61,7 @@ export class AdminScriptImport extends Component {
         <div className={css(styles.buttonDiv)}>
           <RaisedButton
             label="Import"
-            disabled={this.state.importingScript}
+            disabled={this.props.hasPendingJob}
             primary
             onTouchTap={this.startImport}
           />
@@ -99,51 +70,3 @@ export class AdminScriptImport extends Component {
     );
   }
 }
-
-AdminScriptImport.propTypes = {
-  onSubmit: type.func,
-  campaignData: PropTypes.object,
-  mutations: PropTypes.object,
-  pendingJobsData: PropTypes.object
-};
-
-const queries = {
-  pendingJobsData: {
-    query: gql`
-      query getCampaignJobs($campaignId: String!) {
-        campaign(id: $campaignId) {
-          id
-          pendingJobs {
-            id
-            jobType
-            assigned
-            status
-            resultMessage
-          }
-        }
-      }
-    `,
-    options: ownProps => ({
-      variables: {
-        campaignId: ownProps.params.campaignId
-      },
-      pollInterval: 60000 // TODO: revisit
-    })
-  }
-};
-
-const mutations = {
-  importCampaignScript: ownProps => (campaignId, url) => ({
-    mutation: gql`
-      mutation importCampaignScript($campaignId: String!, $url: String!) {
-        importCampaignScript(campaignId: $campaignId, url: $url)
-      }
-    `,
-    variables: {
-      campaignId,
-      url
-    }
-  })
-};
-
-export default loadData({ queries, mutations })(AdminScriptImport);
